Cleanup SubtractMeans
We don't need to convert the type of Image!

+++ utils/augmentations.py
@@ -76,7 +76,7 @@ class SubtractMeans(object):
     def __call__(self, image, boxes=None, labels=None):
         image = image.astype(np.float32)
         image -= self.mean
-        return image.astype(np.float32), boxes, labels
+        return image, boxes, labels
 
 
 class ToAbsoluteCoords(object):
